# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API documentation",
    "version": "1.0.0",
    "description": "API documentation",
    "license": {
      "name": "MIT",
      "url": "https://choosealicense.com/licenses/mit/"
    },
    "contact": {
      "email": "hoangncd3@gmail.com",
      "name": "Hoang Nguyen",
      "url": "https://github.com/khachoangpt"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LogoutPayload": {
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout a shop",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Logout a shop",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutPayload"
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "x-client-id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Shop id"
          },
          {
            "in": "header",
            "name": "x-refresh-token",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Refresh token"
          }
        ],
        "responses": {
          "200": {
            "description": "Logout successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Logout successfully"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}